commonize tester arguments in a single location
& don't allow unknown arguments
